requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
faiss-cpu==1.8.0
numpy==1.24.3
beautifulsoup4==4.12.2
requests==2.31.0
pytest==7.4.3
python-dotenv==1.0.1
httpx==0.26.0
pydantic==2.6.0
aiohttp==3.9.3
scikit-learn==1.3.0
pillow==10.3.0
pytesseract==0.3.10
